breaking-change: only use the adapter logic for file models
Not seeing why we should use the adapter for models other than the file model. The JSON api that is used as default should be sufficient for the others.

Refs: oph-86
